Put ZoteroView Outside Component
for easier reusabillity in other components

Tags: [skip ci]
